authentication/backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
flask-restx
python-dotenv
Flask-Cors
